+++ src/app/[variants]/(main)/(mobile)/me/settings/features/useCategory.tsx
@@ -6,7 +6,7 @@ import { Flexbox } from 'react-layout-kit';
 import urlJoin from 'url-join';
 
 import { CellProps } from '@/components/Cell';
-import { isServerMode } from '@/const/version';
+import { isDeprecatedEdition } from '@/const/version';
 import { SettingsTabs } from '@/store/global/initialState';
 import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
 
@@ -38,17 +38,18 @@ export const useCategory = () => {
         </Flexbox>
       ),
     },
-    showLLM && isServerMode
-      ? {
-          icon: Brain,
-          key: SettingsTabs.Provider,
-          label: t('tab.provider'),
-        }
-      : {
-          icon: Brain,
-          key: SettingsTabs.LLM,
-          label: t('tab.llm'),
-        },
+    showLLM &&
+      (isDeprecatedEdition
+        ? {
+            icon: Brain,
+            key: SettingsTabs.LLM,
+            label: t('tab.llm'),
+          }
+        : {
+            icon: Brain,
+            key: SettingsTabs.Provider,
+            label: t('tab.provider'),
+          }),
     { icon: Mic2, key: SettingsTabs.TTS, label: t('tab.tts') },
     {
       icon: Bot,