+++ src/components/ModelSelect/index.tsx
@@ -2,7 +2,14 @@ import { IconAvatarProps, ModelIcon, ProviderIcon } from '@lobehub/icons';
 import { Avatar, Icon, Tooltip } from '@lobehub/ui';
 import { Typography } from 'antd';
 import { createStyles } from 'antd-style';
-import { Infinity, AtomIcon, LucideEye, LucidePaperclip, ToyBrick } from 'lucide-react';
+import {
+  Infinity,
+  AtomIcon,
+  LucideEye,
+  LucideGlobe,
+  LucidePaperclip,
+  ToyBrick,
+} from 'lucide-react';
 import numeral from 'numeral';
 import { rgba } from 'polished';
 import { FC, memo } from 'react';
@@ -14,7 +21,7 @@ import { AiProviderSourceType } from '@/types/aiProvider';
 import { ChatModelCard } from '@/types/llm';
 import { formatTokenNumber } from '@/utils/format';
 
-const useStyles = createStyles(({ css, token }) => ({
+const useStyles = createStyles(({ css, token, isDarkMode }) => ({
   custom: css`
     width: 36px;
     height: 20px;
@@ -40,6 +47,10 @@ const useStyles = createStyles(({ css, token }) => ({
   tagBlue: css`
     color: ${token.geekblue};
     background: ${token.geekblue1};
+  `,
+  tagCyan: css`
+    color: ${isDarkMode ? token.cyan7 : token.cyan10};
+    background: ${isDarkMode ? token.cyan1 : token.cyan2};
   `,
   tagGreen: css`
     color: ${token.green};
@@ -119,6 +130,17 @@ export const ModelInfoTags = memo<ModelInfoTagsProps>(
           >
             <div className={cx(styles.tag, styles.tagPurple)} style={{ cursor: 'pointer' }}>
               <Icon icon={AtomIcon} />
+            </div>
+          </Tooltip>
+        )}
+        {model.search && (
+          <Tooltip
+            placement={placement}
+            styles={{ root: { pointerEvents: 'none' } }}
+            title={t('ModelSelect.featureTag.search')}
+          >
+            <div className={cx(styles.tag, styles.tagCyan)} style={{ cursor: 'pointer' }} title="">
+              <Icon icon={LucideGlobe} />
             </div>
           </Tooltip>
         )}