+++ src/database/schemas/message.ts
@@ -13,7 +13,7 @@ import {
 import { createSelectSchema } from 'drizzle-zod';
 
 import { idGenerator } from '@/database/utils/idGenerator';
-import { ModelReasoning } from '@/types/message';
+import { GroundingSearch, ModelReasoning } from '@/types/message';
 
 import { timestamps } from './_helpers';
 import { agents } from './agent';
@@ -34,6 +34,8 @@ export const messages = pgTable(
     role: text('role', { enum: ['user', 'system', 'assistant', 'tool'] }).notNull(),
     content: text('content'),
     reasoning: jsonb('reasoning').$type<ModelReasoning>(),
+    search: jsonb('search').$type<GroundingSearch>(),
+    metadata: jsonb('metadata'),
 
     model: text('model'),
     provider: text('provider'),