+++ src/store/chat/slices/aiChat/actions/generateAIChat.ts
@@ -455,7 +455,7 @@ export const generateAIChat: StateCreator<
         await messageService.updateMessageError(messageId, error);
         await refreshMessages();
       },
-      onFinish: async (content, { traceId, observationId, toolCalls, reasoning }) => {
+      onFinish: async (content, { traceId, observationId, toolCalls, reasoning, citations }) => {
         // if there is traceId, update it
         if (traceId) {
           msgTraceId = traceId;
@@ -470,15 +470,26 @@ export const generateAIChat: StateCreator<
         }
 
         // update the content after fetch result
-        await internal_updateMessageContent(
-          messageId,
-          content,
+        await internal_updateMessageContent(messageId, content, {
           toolCalls,
-          !!reasoning ? { content: reasoning, duration } : undefined,
-        );
+          reasoning: !!reasoning ? { content: reasoning, duration } : undefined,
+          search: !!citations ? { citations } : undefined,
+        });
       },
       onMessageHandle: async (chunk) => {
         switch (chunk.type) {
+          case 'citations': {
+            // if there is no citations, then stop
+            if (!chunk.citations || chunk.citations.length <= 0) return;
+
+            internal_dispatchMessage({
+              id: messageId,
+              type: 'updateMessage',
+              value: { search: { citations: chunk.citations } },
+            });
+            break;
+          }
+
           case 'text': {
             output += chunk.text;